feat(emacs): add command for getting source control links
i.e. if you're visiting a file in a repo cloned from a gitlab or a
github origin, get-source-control-link will kill (which in my config
also copies to the system clipboard) a stable link to the location in
code at point.

(I say "stable" because I find it frustrating when links to specific
lines stop working as a branch changes.)